Shiny module server functions for browsing metadata

Handles general browsing, filtering, subsetting of metadata. The module requires two tables: a table of sample ID with all measured features, and a table of measured features and metadata. Generally, users can select specific IDs or a group of IDs and see what features are available through an underlying join of the two input tables.

browseServer(
  id,
  dt_id = NULL,
  index1 = "ID",
  index2 = NULL,
  dt_var = NULL,
  dt_var_index = NULL,
  dt_var_ext = NULL,
  informd1 = system.file("help/ID_list.Rmd", package = "DIVE"),
  informd2 = system.file("help/browse_data.Rmd", package = "DIVE")
)

Arguments

dt_id
Table data by ID.

index1
Name of ID column that serves as primary index.

index2
Optional, name of another column for secondary index.

dt_var
Table data by var/feature.

dt_var_ext
Extended table to provide lookup details of metadata.

informd1
Help content for uploading an ID list.

informd2
Help content for how browsing data works.

input, output, session
Standard shiny boilerplate.