Create a custom cell pack widget based on d3 circlepack.js

A wrapper to produce an html widget representing cells and tissues as circles placed in hierarchy.

cellPackR(
  tab,
  entity = "CellTissue",
  context = "CellTissueContext",
  nested = "|",
  cl = system.file("ontology/cl_2018-11.obo", package = "DIVE"),
  rel_type = "is_a",
  ancestor_classes,
  class_colors,
  corrections = NULL,
  contextcolors
)

Arguments

tab
Table with data for widget.

entity
Name of column containing relevant entity.

context
Name of column containing context.

nested
Deliminater for nested rows, defaulting to "|".