Server module server for matrix with drilldown interaction

Drilldown component with two different modes of disply and linked up to two different inputs

dualDrilldownServer(
  id,
  cdata,
  colorby,
  factorx = function(x) is.factor(cdata[[x]]),
  src1 = reactive(NULL),
  src2 = reactive(NULL)
)

Arguments

cdata
A data.table. This is a reactive to allow the component to be updated with a new table of data.

colorby
A named list where name matches the (factor) variable in cdata to use for color grouping. If there is a named vector, this is passed into scale_color_manual to be used for custom color mapping, e.g. `list(MyGroupingVar = c(A = "red", B = "blue"))` (useful for enforcing a consistent/meaningful color scheme), or `list(MyGroupingVar = NULL)` if custom color scaling isn't necessary. See also details.

factorx
Optional, a function that returns a boolean for whether a variable should be plotted as factor when given the variable name. Useful when for some reason factor variables are numeric or character instead of already factor-encoded. By default, variables are merely checked using base::is.factor, which works when the data is already factor-encoded.

input, output, session
Standard shiny boilerplate.

Details

The drilldown component is, at its simplest, a selectInput through which variables can be selected for adaptive visualization as a boxplot or scatterplot (depending on whether one or two variables are selected, respectively). But optional complexity can be added to the component by making it listen to two other sources; that is, there can be two independent higher-level inputs telling this component what should be selected, analogous to the poor drilldown having two different bosses. Thus the "dual" describes both how the drilldown renders in two different ways as well as being able to be controlled by up to two different sources. Another special thing about this component is that data should actually have a default factor grouping variable to be passed into the aesthetics. If there were no default factor grouping variable, when only one variable is selected we would not be able to generate the boxplot view. It should also be possible to specify multiple variables able to be used for grouping, which the user can switch between, but this isn't implemented yet.