Shiny module server for importing data from GEO

Interactively import GEO data with step-by-step process using a series of modals

getGEOServer(id)

Arguments

id
Character ID for specifying namespace, see shiny::NS.

Value

GEOdata reactive values list object containing $accession, $eset, $pData and $return. $return is an internal status flag for use by multiVCtrlServer to know when a new GEO dataset has been processed completely through the annotation step.

Details

This server function implements interactive GEO data import with three main steps.

Query for accession, extract expression data and check whether of compatible type and format.

Import characteristics data selectively according to user.

Import the feature annotation according to user.