Shiny app server module for multi-views

Assemble various module components into a working one-page application for expression data

multiVServer(
  id,
  hdlist = NULL,
  choices = NULL,
  cdata = NULL,
  preselect = NULL,
  genes = DIVE::gene_symbols
)

Arguments

id
Character ID for specifying namespace, see shiny::NS.

hdlist
A list of matrices representing high dimensional datasets; the names are used for choices.

choices
A list of choices to populate selection menu, defaults to names of datasets in hdlist. When dataset selection should be displayed under different groups, one can pass in a list formatted accordingly for shiny::selectizeInput.

cdata
A data.table of characteristics data, commonly phenotype or clinical data.

preselect
Optional, pre-selected phenotype or clinical variables from cdata.

genes
Gene choices for initialization in shiny::selectizeInput.

Details

First, the server function calls multiVCtrlServer, which returns a named list object which either contains data or NULL. If there is data, a xVServer is dynamically initiated to render data in its own container. If NULL, the function removes the appropriate container using the name of the object, which is something like "i1" and corresponds to the index of the stored global datasets.

While multiVCtrlServer is a global control that controls which datasets are displayed at all, the geneVServer and selectVServer modules are global controls that modify the display of any displayed datasets through applying global filters on attributes that should be present in all or most of the data (for expression matrix data, this means gene/protein and attributes corresponding to samples). Thus, each xVServer component necessarily listens to geneVServer and selectVServer, but each can also have its own indepedent local controls, which takes precedence if enabled.