Shiny server module function for returning a subsetted dataset

Server function for creating a reactive subsetted dataset following user interaction.

refSubsetServer(
  id,
  refdata,
  subsetfeat,
  subsets = NULL,
  datakey = NULL,
  refname = NULL,
  exclude = reactive({ }),
  informd = NULL
)

Arguments

refdata
The reference data.table.

subsetfeat
The name of the column containing the subset factors.

subsets
Optional, a custom selection list for subsets; by default, the list is composed of unique options found in subsetfeat.

exclude
Optional, a list of IDs to exclude.

informd
Optional, relative path to an info Rmarkdown file that can be pulled up in a modal.

input, output, session
Standard shiny boilerplate.

refkey
Optional, a named list containing name/label for creating a key-like column, where the name is the name of the column. See details for intended purpose.

Value

A reactive subsetted data.table.

Details

When many instances of the module are used to create subsets from different datasets (as opposed to different subsets from the same dataset), it might be necessary to use refkey to track the dataset origin of the subset, and especially if the subsets are to be merged on some common column. Generically speaking, datasets can come from different sources/locations, so refkey can be called "Source" and the values might then be S1 and S2.

On the other hand, when this is used to create subsets from the same dataset, an "exclude list" can be passed in to make sure that the subset doesn't contain certain indices/IDs. This is desirable for creating subsets that don't overlap with another or for filtering out rows that don't pass some criteria.