Shiny module server for controlling multi-column view

multiVCtrl is the control hub that provides data and parameters for multiVUI, geneV and selectV.

multiVCtrlServer(
  id,
  cdata,
  hdlist,
  choices = names(hdlist),
  key = "ID",
  checkFun = NULL,
  factorx = NULL,
  preselect = NULL,
  infoRmd = system.file("help/ht_upload.Rmd", package = "DIVE")
)

Arguments

cdata
Data.table of phenotype or clinical data.

hdlist
A list of matrices representing high dimensional datasets; the names are used for choices.

choices
Names referencing datasets in hdlist to be used in selection menu, taken from the names in hdlist by default. However, when dataset selection should be displayed under different groups, one can pass in a list formatted accordingly for selectizeInput.

key
Name of column that contains IDs in cdata that link to samples in hdlist datasets. Defaults to "ID".

checkFun
Optional, a check function used for checking data uploads.

factorx
If not NULL, the given name pattern will be used for recognizing and converting columns in cdata to factor. See details.

preselect
Optional, pre-selected phenotype or clinical variables from cdata.

infoRmd
Optional link to an Rmarkdown document containing description and/or instructions for data upload option.

input, output, session
Standard shiny boilerplate.

Value

A reactive values list containing the data matrix for the parameter

hdata

of the multiV module, as well as parameters for geneV and selectV.

Details

The server logic handles sourcing of large expression datasets with three different methods:

Selecting from available pre-processed datasets.

User-uploaded data.

A beta (least-supported) method of retrieving datasets from GEO.

The module also handles upload of phenotype/clinical data. The module is instantiated with whatever is passed into the parameter cdata and uses a mutable version of cdata that appends user uploads. The data in cdata is supposed to be a phenotype or clinical feature that one usually tries to correlate with expression data and can be numeric or categorical. By default, character columns as well as ones matching factorx in cdata will be converted to factors.